♻️ Refactor: 의존성배열에 데이터 추가, useEffect를 통한 refetch 삭제
