+++ src/app/(received)/received/interesting/page.tsx
@@ -12,17 +12,17 @@ import { Navbar } from '@/shared/ui/Navigation';
 function Page() {
   const handleBack = useHistoryBack();
 
-  const { cards: serverReceivedCards } = useReceivedCardsQuery(null);
+  const { cards: serverReceivedCards, isLoading } = useReceivedCardsQuery(null);
   const { setReceivedCards } = useReceivedCardsStore();
 
   useEffect(() => {
-    setReceivedCards(serverReceivedCards);
-  }, []);
+    if (isLoading) setReceivedCards(serverReceivedCards);
+  }, [isLoading, serverReceivedCards]);
 
   return (
     <div className="flex h-dvh w-full justify-center">
       <div className="flex w-full max-w-[600px] flex-col bg-gray-black">
-        <Appbar page="mypage" title="흥미로운 명함" onLeftClick={handleBack} onRightClick={() => console.log('hi')} />
+        <Appbar page="mypage" title="흥미로운 명함" onLeftClick={handleBack} />
         <div className="flex flex-col gap-4 overflow-y-auto px-5 pb-24 scrollbar-hide">
           {serverReceivedCards.map((value, index) => (
             <ReceivedCard key={index} cardData={value} />
